chore(ajax cache): moved ajax-cache to ajax package
