--- packages/ajax-cache/index.js
@@ -1,5 +1,0 @@
-export {
-  lionCacheRequestInterceptorFactory,
-  lionCacheResponseInterceptorFactory,
-  validateOptions,
-} from './src/ajax-cache.js';
