--- packages/ajax-cache/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@lion/ajax-cache",
-  "version": "0.1.0",
-  "description": "Cache interceptors for @lion/ajax",
-  "license": "MIT",
-  "author": "ing-bank",
-  "homepage": "https://github.com/ing-bank/lion/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ing-bank/lion.git",
-    "directory": "packages/ajax-cache"
-  },
-  "main": "index.js",
-  "module": "index.js",
-  "files": [
-    "*.d.ts",
-    "*.js",
-    "docs",
-    "src",
-    "test",
-    "test-helpers",
-    "translations",
-    "types"
-  ],
-  "scripts": {
-    "debug": "cd ../../ && npm run debug -- --group ajax",
-    "debug:firefox": "cd ../../ && npm run debug:firefox -- --group ajax",
-    "debug:webkit": "cd ../../ && npm run debug:webkit -- --group ajax",
-    "prepublishOnly": "../../scripts/npm-prepublish.js",
-    "test": "cd ../../ && npm run test:browser -- --group ajax"
-  },
-  "sideEffects": false,
-  "keywords": [
-    "ajax",
-    "lion",
-    "web-components"
-  ],
-  "publishConfig": {
-    "access": "public"
-  }
-}
