+++ packages/ajax/index.js
@@ -8,4 +8,10 @@ export {
   addAcceptLanguageHeaderInterceptorFactory,
 } from './src/interceptors.js';
 
+export {
+  lionCacheRequestInterceptorFactory,
+  lionCacheResponseInterceptorFactory,
+  validateOptions,
+} from './src/interceptors-cache.js';
+
 export { jsonPrefixTransformerFactory } from './src/transformers.js';
